Break out logic into CreatedAndUpdated component.
