+++ src/features/import/components/Validation.tsx
@@ -1,10 +1,10 @@
-import { Box, Stack } from '@mui/system';
+import { Box, Stack, Typography } from '@mui/material';
 import { FC, useState } from 'react';
-import { Typography, useTheme } from '@mui/material';
 
 import AddedOrgs from './AddedOrgs';
 import AddedTags from './AddedTags';
 import ChangedFields from './ChangedFields';
+import CreatedAndUpdated from './CreatedAndUpdated';
 import messageIds from 'features/import/l10n/messageIds';
 import { Msg } from 'core/i18n';
 import useAlerts from 'features/import/hooks/useAlerts';
@@ -83,7 +83,6 @@ const Validation: FC<ValidationProps> = ({ onClickBack, onDisabled }) => {
       },
     },
   };
-  const theme = useTheme();
   const [checkedIndexes, setCheckedIndexes] = useState<number[]>([]);
   const { orgId } = useNumericRouteParams();
   const { numPeopleWithOrgsAdded, orgsWithNewPeople } = useOrgUpdates(
@@ -122,56 +121,7 @@ const Validation: FC<ValidationProps> = ({ onClickBack, onDisabled }) => {
           height="100%"
           sx={{ mt: 2, overflowY: 'scroll' }}
         >
-          <Stack direction="row" spacing={2}>
-            <Box
-              border={1}
-              borderColor={theme.palette.grey[300]}
-              borderRadius={1}
-              padding={2}
-              width="100%"
-            >
-              <Msg
-                id={messageIds.validation.updateOverview.created}
-                values={{
-                  numPeople: fake.summary.peopleCreated.total,
-                  number: (
-                    <Typography
-                      sx={{
-                        color: theme.palette.success.main,
-                      }}
-                      variant="h2"
-                    >
-                      {fake.summary.peopleCreated.total}
-                    </Typography>
-                  ),
-                }}
-              />
-            </Box>
-            <Box
-              border={1}
-              borderColor={theme.palette.grey[300]}
-              borderRadius={1}
-              padding={2}
-              width="100%"
-            >
-              <Msg
-                id={messageIds.validation.updateOverview.updated}
-                values={{
-                  numPeople: fake.summary.peopleUpdated.total,
-                  number: (
-                    <Typography
-                      sx={{
-                        color: theme.palette.info.light,
-                      }}
-                      variant="h2"
-                    >
-                      {fake.summary.peopleUpdated.total}
-                    </Typography>
-                  ),
-                }}
-              />
-            </Box>
-          </Stack>
+          <CreatedAndUpdated summary={fake.summary} />
           <Stack spacing={2} sx={{ mt: 2 }}>
             <ChangedFields
               changedFields={fake.summary.peopleUpdated.byField}
